requirements.txt
# Core Dependencies
sqlalchemy==2.0.23
fastapi>=0.100.0
uvicorn>=0.22.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
PyYAML>=6.0.1
redis>=5.0.0
pydantic>=2.0.0
alembic
torch
httpx
numpy
apscheduler